Engine: move "logs" tail logic into the log driver
Move logic for handling the --tail option for the "logs" endpoint into
the log driver itself.  This means that each driver has to implement the
"get the last N items" logic itself, but the benefit is that GetReader()
can start to return Reader objects, and we don't silently require that
they also be os.File objects.

Also pass the log cutoff date as a hint to the log driver, so that the
log driver can take advantage of that information if there's a benefit
to its doing so.  Leave logic for filtering using that date in place in
the main logs handler, because it doesn't cost us much, and so that the
log driver can avoid dealing with it if it wants to.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com> (github: nalind)
Signed-off-by: Dan Walsh <dwalsh@redhat.com> (github: rhatdan)